pyproject.toml
[project]
name = "agent-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "fastapi>=0.116.0",
    "fitz>=0.0.1.dev2",
    "fonttools>=4.58.5",
    "googletrans>=4.0.2",
    "openai>=1.93.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psutil>=7.0.0",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.2",
]
